# openapi.yaml
openapi: 3.0.1
info:
  title: AKIRAStack
  description: >-
    This is the API Documentation for the AkiraStack free check score
  version: 1.0.0
servers:
  - url: 'https://continuous-security.secops.neuralcompany.team/api/security'
components:
  securitySchemes:
    bearer:            
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /v1/git-repos:
    get:
      tags:
      - GitRepos
      operationId: GitRepos
      summary: Get list of Git repos of the user
      responses:
        '200':
          description: connection successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                  repositories_count:
                    type: integer
                  repositories_data:
                    type: array
                    items:
                     properties:
                      full_name:
                       type: string
                      private:
                       type: string
                      html_url:
                       type: string
                      clone_url:
                       type: string
                      name:
                       type: string
        '500':
          description: connection failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                  message:
                    type: string
      security:
        - bearer: []
...